requirements.txt
langgraph>=0.0.15
langchain-core>=0.1.17
langchain-community>=0.0.19
langchain-anthropic>=0.2.1
langchain-together>=0.0.2.post1
langchain_google_genai
langchain_mistralai
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioresponses==0.7.8
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autopep8==2.3.2
beautifulsoup4==4.13.5
beautifulsoup4==4.13.5
black==25.1.0
bs4==0.0.2
bs4==0.0.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
coverage==7.10.2
deptry==0.23.1
distro==1.9.0
elevenlabs==2.8.1
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-mcp-adapters==0.1.9
langchain-mistralai==0.2.11
langchain-openai==0.3.28
langchain-together==0.3.1
langsmith==0.4.10
mccabe==0.7.0
mcp==1.12.3
multidict==6.6.3
mutagen==1.47.0
mypy_extensions==1.1.0
numpy==2.3.2
numpy==2.3.2
oauthlib==3.3.1
obsws-python==1.8.0
openai==1.98.0
orjson==3.11.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dotenv==1.1.1
python-multipart==0.0.20
python-statemachine==2.5.0
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.13.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.7
schedule==1.2.2
SClib==1.0.0
schedule==1.2.2
SClib==1.0.0
sniffio==1.3.1
soundcloud==0.5.0
soundcloud-lib==0.6.1
soupsieve==2.8
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
